Refactor wormhole utils
