+++ wormhole-kernel/wormhole-metadata/src/main/java/com/zergclan/wormhole/metadata/plan/filter/precise/convertor/NodeNameConvertorMetaData.java
@@ -18,7 +18,7 @@
 package com.zergclan.wormhole.metadata.plan.filter.precise.convertor;
 
 import com.zergclan.wormhole.common.constant.MarkConstant;
-import com.zergclan.wormhole.common.util.PropertiesExtractor;
+import com.zergclan.wormhole.common.util.ValueExtractor;
 import com.zergclan.wormhole.metadata.plan.filter.FilterMetaData;
 import com.zergclan.wormhole.metadata.plan.filter.FilterType;
 import lombok.Getter;
@@ -62,8 +62,8 @@ public final class NodeNameConvertorMetaData implements FilterMetaData {
      * @return {@link NodeNameConvertorMetaData}
      */
     public static NodeNameConvertorMetaData builder(final String taskIdentifier, final int order, final Properties props) {
-        String sourceName = PropertiesExtractor.extractRequiredString(props, "sourceName");
-        String targetName = PropertiesExtractor.extractRequiredString(props, "targetName");
+        String sourceName = ValueExtractor.extractRequiredString(props, "sourceName");
+        String targetName = ValueExtractor.extractRequiredString(props, "targetName");
         return new FilterBuilder(taskIdentifier, order, sourceName, targetName).build();
     }
