+++ wormhole-kernel/wormhole-metadata/src/main/java/com/zergclan/wormhole/metadata/plan/filter/precise/convertor/PatternedDataTimeConvertorMetaData.java
@@ -18,7 +18,7 @@
 package com.zergclan.wormhole.metadata.plan.filter.precise.convertor;
 
 import com.zergclan.wormhole.common.constant.MarkConstant;
-import com.zergclan.wormhole.common.util.PropertiesExtractor;
+import com.zergclan.wormhole.common.util.ValueExtractor;
 import com.zergclan.wormhole.metadata.plan.filter.FilterMetaData;
 import com.zergclan.wormhole.metadata.plan.filter.FilterType;
 import lombok.Getter;
@@ -64,9 +64,9 @@ public final class PatternedDataTimeConvertorMetaData implements FilterMetaData 
      * @return {@link PatternedDataTimeConvertorMetaData}
      */
     public static PatternedDataTimeConvertorMetaData builder(final String taskIdentifier, final int order, final Properties props) {
-        String sourceName = PropertiesExtractor.extractRequiredString(props, "sourceName");
-        String sourcePattern = PropertiesExtractor.extractRequiredString(props, "sourcePattern");
-        String targetPattern = PropertiesExtractor.extractRequiredString(props, "targetPattern");
+        String sourceName = ValueExtractor.extractRequiredString(props, "sourceName");
+        String sourcePattern = ValueExtractor.extractRequiredString(props, "sourcePattern");
+        String targetPattern = ValueExtractor.extractRequiredString(props, "targetPattern");
         return new PatternedDataTimeConvertorMetaData.FilterBuilder(taskIdentifier, order, sourceName, sourcePattern, targetPattern).build();
     }
