+++ wormhole-kernel/wormhole-metadata/src/main/java/com/zergclan/wormhole/metadata/plan/filter/precise/editor/FixedNodeEditorMetaData.java
@@ -18,7 +18,7 @@
 package com.zergclan.wormhole.metadata.plan.filter.precise.editor;
 
 import com.zergclan.wormhole.common.constant.MarkConstant;
-import com.zergclan.wormhole.common.util.PropertiesExtractor;
+import com.zergclan.wormhole.common.util.ValueExtractor;
 import com.zergclan.wormhole.metadata.plan.filter.FilterMetaData;
 import com.zergclan.wormhole.metadata.plan.filter.FilterType;
 import com.zergclan.wormhole.metadata.plan.node.DataNodeTypeMetaData;
@@ -66,9 +66,9 @@ public final class FixedNodeEditorMetaData implements FilterMetaData {
      * @return {@link FixedNodeEditorMetaData}
      */
     public static FixedNodeEditorMetaData builder(final String taskIdentifier, final int order, final Properties props) {
-        String sourceName = PropertiesExtractor.extractRequiredString(props, "sourceName");
-        String defaultValue = PropertiesExtractor.extractRequiredString(props, "defaultValue");
-        String dataType = PropertiesExtractor.extractRequiredString(props, "defaultValue");
+        String sourceName = ValueExtractor.extractRequiredString(props, "sourceName");
+        String defaultValue = ValueExtractor.extractRequiredString(props, "defaultValue");
+        String dataType = ValueExtractor.extractRequiredString(props, "defaultValue");
         return new FilterBuilder(taskIdentifier, order, sourceName, defaultValue, DataNodeTypeMetaData.DataType.valueOf(dataType.toUpperCase(Locale.ROOT))).build();
     }
